src/serving/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
evidently>=0.4.10
onnxruntime>=1.16.0
numpy>=1.24.0
pandas>=2.0.0
feast[redis]>=0.35.0
azure-ai-ml>=1.12.0
azure-identity>=1.15.0
prometheus-client>=0.19.0
python-dotenv>=1.0.0
requests>=2.31.0
